StudyTracker/backend/requirements.txt
python-dotenv~=1.0.1
langgraph~=0.2.26
langchain-core~=0.2.41
langchain-community~=0.2.17
langchain-groq
langchain~=0.2.16
crewai~=0.63.2
crewai[tools]
Flask~=3.0.3
Flask-Cors~=5.0.0
duckduckgo-search